src/jupyrest/pyproject.toml
[tool.poetry]
name = "jupyrest"
version = "0.1.0"
description = "A tool to expose Jupyter notebooks as a REST API."
authors = ["Koushik Krishnan <kokrishn@microsoft.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
jsonschema = "^3.2.0"
typing-extensions = { version = "^3.7.4", python = "<3.8" }
nbformat = "^5.1.3"
papermill = "~2.2"
nbclient = "^0.5.1"
entrypoints = "^0.3"
scrapbook = "^0.5.0"
pydantic = "^1.8"
grpcio = "^1.38.0"
fastapi = "^0.65.2"
grpcio-tools = "^1.38.0"
jupyter-client = "<=6.1.12"
textwrap3 = "^0.9.2"
click = "^7.1.2"
typer = "^0.3.2"
uvicorn = "^0.14.0"
pywin32 = { version = "==301", markers = "sys_platform == 'win32'" }
setuptools-rust = "^0.12.1"
ipython = "^7.25.0"
ipykernel = "^6.0.1"
azure-storage-blob = "^12.8.1"
black = "==21.7b0"
cryptography = "<=3.4.7"
opentelemetry-api = "^1.5.0"
protobuf = "3.20.*"
ipython-genutils = "^0.2.0"
nbconvert = "^7.2.10"
azure-functions = "^1.13.3"
aiofiles = "^23.2.1"
jupyter-book = "^0.15.1"
importlib-metadata = "<5.0"
anyio = "^4.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
mypy-protobuf = "^2.4"
pytest-asyncio = "^0.15.1"
grpcio-testing = "^1.37.1"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

